+++ lib/dead_end/who_dis_syntax_error.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module SyntaxErrorSearch
+module DeadEnd
   # Determines what type of syntax error is in the source
   #
   # Example: