+++ spec/unit/dead_end_spec.rb
@@ -2,10 +2,10 @@
 
 require_relative "../spec_helper.rb"
 
-module SyntaxErrorSearch
-  RSpec.describe SyntaxErrorSearch do
+module DeadEnd
+  RSpec.describe DeadEnd do
     it "has a version number" do
-      expect(SyntaxErrorSearch::VERSION).not_to be nil
+      expect(DeadEnd::VERSION).not_to be nil
     end
 
     it "detects require error and adds a message with auto mode" do
@@ -30,10 +30,16 @@ module SyntaxErrorSearch
           require_relative "./script.rb"
         EOM
 
-        out = `ruby -I#{lib_dir} -rsyntax_search/auto #{require_rb} 2>&1`
+        out = `ruby -I#{lib_dir} -rdead_end/auto #{require_rb} 2>&1`
 
         expect(out).to include("Unmatched `end` detected")
-        expect(out).to include("Run `$ syntax_search")
+        expect(out).to include("Run `$ dead_end")
+        expect($?.success?).to be_falsey
+
+        out = `ruby -I#{lib_dir} -rdead_end #{require_rb} 2>&1`
+
+        expect(out).to include("Unmatched `end` detected")
+        expect(out).to include("Run `$ dead_end")
         expect($?.success?).to be_falsey
       end
     end
@@ -60,10 +66,10 @@ module SyntaxErrorSearch
           require_relative "./script.rb"
         EOM
 
-        out = `ruby -I#{lib_dir} -rsyntax_search/fyi #{require_rb} 2>&1`
+        out = `ruby -I#{lib_dir} -rdead_end/fyi #{require_rb} 2>&1`
 
         expect(out).to_not include("This code has an unmatched")
-        expect(out).to include("Run `$ syntax_search")
+        expect(out).to include("Run `$ dead_end")
         expect($?.success?).to be_falsey
       end
     end